Add `dropLast.js` and `identical.js`
