+++ SIZES.md
@@ -55,7 +55,7 @@
 | dissocPath | 136 B | 708 B | -572 B |
 | divide | 82 B | 230 B | -148 B |
 | drop | 139 B | 782 B | -643 B |
-| dropLast | 142 B | 983 B | -841 B |
+| dropLast | 149 B | 983 B | -834 B |
 | dropLastWhile | 141 B | 1443 B | -1302 B |
 | dropRepeats | 375 B | 1845 B | -1470 B |
 | dropRepeatsWith | 119 B | 722 B | -603 B |
@@ -88,18 +88,18 @@
 | has | 85 B | 240 B | -155 B |
 | hasIn | 85 B | 231 B | -146 B |
 | head | 31 B | 301 B | -270 B |
-| identical | 109 B | 249 B | -140 B |
+| identical | 121 B | 249 B | -128 B |
 | identity | 29 B | 150 B | -121 B |
 | ifElse | 110 B | 687 B | -577 B |
 | inc | 32 B | 250 B | -218 B |
 | indexBy | 98 B | 1438 B | -1340 B |
 | indexed | 40 B | n/a B | n/a B |
 | indexOf | 110 B | 1570 B | -1460 B |
-| init | 152 B | 493 B | -341 B |
+| init | 109 B | 493 B | -384 B |
 | innerJoin | 155 B | 439 B | -284 B |
 | insert | 124 B | 378 B | -254 B |
 | insertAll | 132 B | 379 B | -247 B |
-| intersection | 515 B | 2360 B | -1845 B |
+| intersection | 511 B | 2360 B | -1849 B |
 | intersperse | 117 B | 409 B | -292 B |
 | invert | 58 B | 534 B | -476 B |
 | invertObj | 44 B | 520 B | -476 B |
