+++ lib/dropLast/dropLast.js
@@ -1,0 +1,11 @@
+import toArray from '../toArray'
+
+export default function dropLast(n, from) {
+  return n <= 0
+    ? from
+    : typeof from === 'string'
+      ? toArray(from)
+        .slice(0, -n)
+        .join('')
+      : toArray(from).slice(0, -n)
+}
