+++ lib/dropLast/index.js
@@ -1,12 +1,4 @@
 import curry2 from '../_internal/_curry2'
-import toArray from '../toArray'
+import dropLast from './dropLast'
 
-export default curry2(function dropLast (n, from) {
-  return n <= 0
-    ? from
-    : typeof from === 'string'
-      ? toArray(from)
-        .slice(0, -n)
-        .join('')
-      : toArray(from).slice(0, -n)
-})
+export default curry2(dropLast)
