+++ lib/init/index.js
@@ -1,3 +1,5 @@
-import dropLast from '../dropLast'
+import dropLast from '../dropLast/dropLast'
 
-export default dropLast(1)
+export default function init(arr) {
+  return dropLast(1, arr)
+}
