+++ lib/intersection/index.js
@@ -1,6 +1,6 @@
 import _curry2 from '../_internal/_curry2'
 import equals from '../equals'
-import identical from '../identical'
+import identical from '../identical/identical'
 
 export default _curry2(function intersection(list1, list2) {
   var resObj = {}
